refactor: move generationInfo building from buildPrompt to sendChat
